Split settings validator in separate validators for sub-settings
* refs #217
